requirements.txt
annotated-types==0.7.0
anyio==4.9.0
APScheduler==3.11.0
beautifulsoup4==4.13.4
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
fastapi==0.115.12
feedparser==6.0.11
frozendict==2.4.6
h11==0.16.0
idna==3.10
multitasking==0.0.11
numpy==2.2.5
pandas==2.2.3
peewee==3.17.9
platformdirs==4.3.7
pydantic==2.11.3
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
requests==2.32.3
sgmllib3k==1.0.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.7
starlette==0.46.2
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
tzlocal==5.3.1
urllib3==2.4.0
uvicorn==0.34.2
yfinance==0.2.56
